refactor: menu test 패키지 분리
